Minor changes on unit tests

+++ src/main/java/com/musala/gateway/service/GatewayService.java
@@ -8,5 +8,5 @@ public interface GatewayService {
 
     void save(GatewayAddDto gatewayAddDto);
 
-    void printInfoForAGateway(Integer id);
+    void printInfoForAGateway(long id);
 }
